Cell tap opens character details
Again, this page has nothing but a label with the character's name.

A Navigator was added to take navigation logic out of the view
controllers and don't let them know each other. I've done a very dirty
hack with that static Navigator instance, but at least it is private and
allows me to inject logic into the VCs - they don't even need to know
the Navigator itself. I'm not happy with this approach though, help me
code reviewers! :D