StoryType can be any String
There seems to be no closed set to define story types, in fact it can
even be an empty string or inconsistently typed like "pinup" and
"pin-up", so the app won't limit it anymore to a known set of types.

Also added "png" to the known set of image extensions.